TPP-1787: ads async reports refactored and working in v3 and v5
